ISSUE FIXED
According to my RnD, the previously implemented "ON DUPLICATE" works
only with MySQL DB (this was probably the DB used when developing the
plugin).
The updated "ON CONFLICT" query works in all major DBs, except for
MySQL.

However, the default DB for OpenFire (HSQL DB) supports neither of them.

We can either change the query according to the DB(as suggested), or
provide this implementation.

Try/Catch or a case based SQL approach does not wok in DML of any SQL,
hence a count query would be more appropriate.

Also, I do not know kotlin and would very much appreciate any input in
optimizing the code if possible.

Note: With this approach, we can effectively remove the unique constraint on the table as it would merely be a precaution with time overhead.